+++ src/transactions/test/mocks/MockLedgerDelta.h
@@ -17,6 +17,7 @@ class MockLedgerDelta : public LedgerDelta
     MOCK_METHOD1(deleteEntry, void(LedgerKey const& key));
     MOCK_METHOD1(modEntry, void(EntryFrame const& entry));
     MOCK_METHOD1(recordEntry, void(EntryFrame const& entry));
+    MOCK_METHOD0(deleteEntryDuplicate, void());
     MOCK_METHOD1(mergeEntries, void(LedgerDelta& other));
     MOCK_METHOD0(commit, void());
     MOCK_METHOD0(rollback, void());